ANPR/backend/requirements.txt
<<<<<<< HEAD
ultralytics
torch
opencv-python
easyocr
matplotlib
numpy
Pillow
PyQt6
pymongo
cloudinary
loguru
python-dotenv
levensthein
streamlit
python-multipart
openpyxl
fastapi
deep-sort-realtime
=======
cloudinary==1.44.1
deep_sort_realtime==1.3.2
easyocr==1.7.2
fastapi==0.115.13
loguru==0.7.3
numpy==2.3.1
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
pandas==2.3.0
pymongo==4.13.2
python-dotenv==1.1.1
python_Levenshtein==0.27.1
torch==2.7.1
ultralytics==8.3.155
>>>>>>> f5520ec (ANPR updated)
